+++ backend/src/routes/employees.ts
@@ -32,7 +32,7 @@ router.get("/includeUnregistered", async (req: Request, res: Response) => {
         if (findIndex === -1) {
           const findEmployeeIndex = employees.findIndex(f => {
             let name = `${f['firstName']}`;
-            if(f['middleName']) name += ` ${f['middleName']}`
+            if (f['middleName']) name += ` ${f['middleName']}`
             name += ` ${f['lastName']}`;
 
             return f['code'] === value || name === value
@@ -134,9 +134,15 @@ router.post("/", [
     }
     try {
       const token = req.cookies.auth_token;
-      const decodedToken: any = jwt.verify(token, process.env.JWT_SECRET_KEY as string);
+      const decodedToken: any = await fetch(`${process.env.ROCKS_DEV_API_URL}/auth/check`, {
+        method: "POST",
+        credentials: "include",
+        headers: {
+          "Content-Type": "application/json"
+        },
+      });
 
-      if (decodedToken.role !== "ADMIN") {
+      if (false) { // TODO: UPDATE WHEN ROCKS API IS UPDATED WITH USER ROLES
         return res.status(403).json({ message: "Only users with admin role can perform this action" });
       }
 
@@ -201,9 +207,15 @@ router.put("/history/:code",
     }
     try {
       const token = req.cookies.auth_token;
-      const decodedToken: any = jwt.verify(token, process.env.JWT_SECRET_KEY as string);
+      const decodedToken: any = await fetch(`${process.env.ROCKS_DEV_API_URL}/auth/check`, {
+        method: "POST",
+        credentials: "include",
+        headers: {
+          "Content-Type": "application/json"
+        },
+      });
 
-      if (decodedToken.role !== "ADMIN") {
+      if (false) { // TODO: UPDATE WHEN ROCKS API IS UPDATED WITH USER ROLES
         return res.status(403).json({ message: "Only users with admin role can perform this action" });
       }
 
@@ -255,9 +267,15 @@ router.put("/:code", [
     }
     try {
       const token = req.cookies.auth_token;
-      const decodedToken: any = jwt.verify(token, process.env.JWT_SECRET_KEY as string);
+      const decodedToken: any = await fetch(`${process.env.ROCKS_DEV_API_URL}/auth/check`, {
+        method: "POST",
+        credentials: "include",
+        headers: {
+          "Content-Type": "application/json"
+        },
+      });
 
-      if (decodedToken.role !== "ADMIN") {
+      if (false) { // TODO: UPDATE WHEN ROCKS API IS UPDATED WITH USER ROLES
         return res.status(403).json({ message: "Only users with admin role can perform this action" });
       }
 
@@ -334,9 +352,15 @@ router.put("/:code", [
 router.delete("/:code", verifyToken, async (req: Request, res: Response) => {
   try {
     const token = req.cookies.auth_token;
-    const decodedToken: any = jwt.verify(token, process.env.JWT_SECRET_KEY as string);
+    const decodedToken: any = await fetch(`${process.env.ROCKS_DEV_API_URL}/auth/check`, {
+      method: "POST",
+      credentials: "include",
+      headers: {
+        "Content-Type": "application/json"
+      },
+    });
 
-    if (decodedToken.role !== "ADMIN") {
+    if (false) { // TODO: UPDATE WHEN ROCKS API IS UPDATED WITH USER ROLES
       return res.status(403).json({ message: "Only users with admin role can perform this action" });
     }