+++ frontend/src/components/ui/alert.tsx
@@ -4,7 +4,7 @@ import { cva, type VariantProps } from "class-variance-authority"
 import { cn } from "@/lib/utils"
 
 const alertVariants = cva(
-  "relative w-full rounded-lg border p-4 [&>svg~*]:pl-7 [&>svg+div]:translate-y-[-3px] [&>svg]:absolute [&>svg]:left-4 [&>svg]:top-4 [&>svg]:text-foreground",
+  "relative w-full rounded-lg border p-4",
   {
     variants: {
       variant: {
@@ -18,8 +18,9 @@ const alertVariants = cva(
         warningOutline:
           "border-warning/50 text-warning dark:border-warning [&>svg]:text-warning",
         success: 
-          "border-primary bg-primary/10 text-primary"
-
+          "border-primary bg-primary/10 text-primary",
+        error:
+          "border-none text-error bg-error-bg"
         },
 
     },
@@ -60,7 +61,7 @@ const AlertDescription = React.forwardRef<
 >(({ className, ...props }, ref) => (
   <div
     ref={ref}
-    className={cn("text-sm [&_p]:leading-relaxed", className)}
+    className={cn("text-xs [&_p]:leading-relaxed", className)}
     {...props}
   />
 ))