+++ frontend/src/components/ui/input.tsx
@@ -1,5 +1,6 @@
 import * as React from "react"
 
+
 import { cn } from "@/lib/utils"
 
 export interface InputProps
@@ -7,11 +8,13 @@ export interface InputProps
 
 const Input = React.forwardRef<HTMLInputElement, InputProps>(
   ({ className, type, ...props }, ref) => {
+    const { "aria-invalid": error } = props; // Hacky way to do it, but it works
+
     return (
       <input
         type={type}
         className={cn(
-          "flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50",
+          `flex h-10 w-full rounded-md border border-input px-3 py-2 text-sm text-black ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${error === true || error === 'true' ? 'border-error shadow-[0px_0px_2px_2px_rgba(243,53,53,0.3)]' : ''}`,
           className
         )}
         ref={ref}