+++ frontend/tailwind.config.cjs
@@ -15,25 +15,46 @@ module.exports = {
       center: true,
       padding: "2rem",
       screens: {
+        "3xl": "1536px",
         "2xl": "1400px",
+        "tablet": "1024px",
+        "mobile-sm": "320px",
+        "mobile-lg": "640px"
       },
     },
     extend: {
       colors: {
-        border: "hsl(var(--border))",
+        body: "var(--body-new)",
         input: "hsl(var(--input))",
         ring: "hsl(var(--ring))",
         highlight: "hsl(var(--highlight))",
         darker: "hsl(var(--darker))",
-        background: "hsl(var(--background))",
+        error: {
+          DEFAULT: "var(--error)",
+          bg: "var(--error-bg)"
+        },
+        background: {
+          DEFAULT: "hsl(var(--background))",
+          new: "var(--background-new)"
+        },
         foreground: "hsl(var(--foreground))",
+        "fs-green": {
+          DEFAULT: "var(--fs-green)",
+          hover: "var(--fs-green-hover)"
+        },
+        border: {
+          DEFAULT: "hsl(var(--border))",
+          new: "var(--border-new)"
+        },
         primary: {
           DEFAULT: "hsl(var(--primary))",
           foreground: "hsl(var(--primary-foreground))",
+          new: "var(--primary-new)"
         },
         secondary: {
           DEFAULT: "hsl(var(--secondary))",
           foreground: "hsl(var(--secondary-foreground))",
+          new: "var(--secondary-new)"
         },
         destructive: {
           DEFAULT: "hsl(var(--destructive))",
@@ -46,6 +67,7 @@ module.exports = {
         muted: {
           DEFAULT: "hsl(var(--muted))",
           foreground: "hsl(var(--muted-foreground))",
+          new: "var(--muted-new)"
         },
         accent: {
           DEFAULT: "hsl(var(--accent))",
@@ -78,6 +100,19 @@ module.exports = {
       animation: {
         "accordion-down": "accordion-down 0.2s ease-out",
         "accordion-up": "accordion-up 0.2s ease-out",
+      },
+      backgroundImage: {
+        'login-bg': "linear-gradient(164.19deg, rgba(36, 67, 66, 0.4) 11.03%, #244342 88.77%), url('/src/assets/laptop-background.jpg')"
+      },
+      screens: {
+        "3xl": "1536px",
+        "2xl": "1400px",
+        "tablet": "1024px",
+        "mobile-sm": "320px",
+        "mobile-lg": "640px"
+      },
+      fontFamily: {
+        poppins: ["Poppins", "sans-serif"]
       },
     },
   },